+++ bin/firebase_verify_token_dart.dart
@@ -1,10 +1,25 @@
 import 'package:firebase_verify_token_dart/firebase_verify_token_dart.dart';
 
 void main() async {
-  FirebaseVerifyToken.projectId = '';
+  FirebaseVerifyToken.projectIds = ['test', 'test-1'];
 
-  final result = await FirebaseVerifyToken.verify('token');
-
+  final result = await FirebaseVerifyToken.verify(
+    'token',
+    onVerifySuccessful: ({required bool status, String? projectId}) {
+      print('STATUS: $status');
+      print('PROJECT ID: $projectId');
+    },
+  );
+  await FirebaseVerifyToken.verify(
+    'my-token-string',
+    onVerifySuccessful: ({required bool status, String? projectId}) {
+      if (status) {
+        print('Token verified for project: $projectId');
+      } else {
+        print('Token verification failed');
+      }
+    },
+  );
   if (result) {
     print('TOKEN IS VERIFIED');
   } else {