+++ CHANGELOG.md
@@ -1,4 +1,16 @@
-## [0.0.7]
+# Changelog
+
+## 1.0.0
+
+- Replaced `projectId` with `projectIds` to support multi-project verification.
+- Added the `onVerifySuccessful` callback function to execute code once the token is verified.
+- Reduced verification times by caching Google `kids` and `NTP`.
+- Updated dependencies.
+- Optimized parsing algorithms.
+- Clean code.
+- Updated documentation.
+
+## 0.0.7
 
 - Publish package